backend/requirements.txt
fastapi
uvicorn
sqlmodel
psycopg2-binary
pgvector
python-jose[cryptography]
bcrypt==4.0.1
passlib[bcrypt]
python-multipart
langchain
langchain-openai
python-dotenv
httpx
pydantic-settings
openai>=1.3.0
pypdf>=3.17.0
PyMuPDF>=1.23.0
python-docx
openpyxl
celery>=5.3.0
redis>=5.0.0
boto3>=1.29.0
slowapi
alembic>=1.12.0
prometheus-client
sentence-transformers
pytesseract
Pillow
beautifulsoup4
requests
lxml
